# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Naruto RPG Community API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  articleCount?: number;
}

export interface ArticleSummary {
  id: number;
  title: string;
  slug: string;
  categoryId: number;
  categoryName: string;
  excerpt: string;
  updatedAt: string;
}

export interface Article {
  id: number;
  title: string;
  slug: string;
  categoryId: number;
  categoryName: string;
  parentId?: number | null;
  content: string;
  excerpt: string;
  updatedAt: string;
  createdAt: string;
}

export interface ArticleInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  slug: string;
  categoryId: number;
  parentId?: number | null;
  content: string;
  excerpt: string;
}

export interface ArticleUpdate {
  /** @minLength 1 */
  title?: string;
  slug?: string;
  categoryId?: number;
  parentId?: number | null;
  content?: string;
  excerpt?: string;
}

export interface Section {
  id: number;
  title: string;
  slug: string;
  description: string;
  link: string;
  icon: string;
  visible: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface SectionInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  slug: string;
  description?: string;
  link?: string;
  icon?: string;
  visible?: boolean;
  sortOrder?: number;
}

export interface SectionUpdate {
  /** @minLength 1 */
  title?: string;
  slug?: string;
  description?: string;
  link?: string;
  icon?: string;
  visible?: boolean;
  sortOrder?: number;
}

export interface SectionOrder {
  /** Array of section IDs in desired order */
  order: number[];
}

export interface LoreChapter {
  id: number;
  title: string;
  content: string;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface LoreChapterInput {
  /** @minLength 1 */
  title: string;
  content?: string;
  sortOrder?: number;
}

export interface LoreChapterUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  sortOrder?: number;
}

export interface ChapterOrder {
  /** Array of chapter IDs in desired order */
  order: number[];
}

export type WikiStatsCategoryBreakdownItem = {
  categoryName: string;
  count: number;
};

export interface WikiStats {
  totalArticles: number;
  totalCategories: number;
  categoryBreakdown: WikiStatsCategoryBreakdownItem[];
}

export type ListArticlesParams = {
category?: string;
search?: string;
};

export type ListSectionsParams = {
visibleOnly?: boolean;
};