# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Naruto RPG Community API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: articles
    description: Wiki articles
  - name: categories
    description: Wiki categories
  - name: sections
    description: Home page sections
  - name: lore-chapters
    description: Lore page chapters
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List articles, optionally filtered by category
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArticleSummary"
    post:
      operationId: createArticle
      tags: [articles]
      summary: Create a new article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"

  /articles/recent:
    get:
      operationId: listRecentArticles
      tags: [articles]
      summary: Get recently updated articles
      responses:
        "200":
          description: Recent articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArticleSummary"

  /articles/stats:
    get:
      operationId: getWikiStats
      tags: [articles]
      summary: Get wiki statistics (totals per category)
      responses:
        "200":
          description: Wiki stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WikiStats"

  /articles/{id}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get a single article with full content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
    patch:
      operationId: updateArticle
      tags: [articles]
      summary: Update an article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleUpdate"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
    delete:
      operationId: deleteArticle
      tags: [articles]
      summary: Delete an article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all wiki categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /sections:
    get:
      operationId: listSections
      tags: [sections]
      summary: List sections, optionally only visible ones
      parameters:
        - name: visibleOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Sections list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Section"
    post:
      operationId: createSection
      tags: [sections]
      summary: Create a new section
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"

  /sections/reorder:
    patch:
      operationId: reorderSections
      tags: [sections]
      summary: Update sort order for multiple sections
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionOrder"
      responses:
        "200":
          description: Sections reordered
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Section"

  /sections/{id}:
    patch:
      operationId: updateSection
      tags: [sections]
      summary: Update a section
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionUpdate"
      responses:
        "200":
          description: Updated section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
        "404":
          description: Not found
    delete:
      operationId: deleteSection
      tags: [sections]
      summary: Delete a section
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /lore-chapters:
    get:
      operationId: listLoreChapters
      tags: [lore-chapters]
      summary: List all lore chapters
      responses:
        "200":
          description: List of chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoreChapter"
    post:
      operationId: createLoreChapter
      tags: [lore-chapters]
      summary: Create a new lore chapter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoreChapterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoreChapter"

  /lore-chapters/reorder:
    patch:
      operationId: reorderLoreChapters
      tags: [lore-chapters]
      summary: Update sort order for chapters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterOrder"
      responses:
        "200":
          description: Chapters reordered
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoreChapter"

  /lore-chapters/{id}:
    patch:
      operationId: updateLoreChapter
      tags: [lore-chapters]
      summary: Update a chapter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoreChapterUpdate"
      responses:
        "200":
          description: Updated chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoreChapter"
        "404":
          description: Not found
    delete:
      operationId: deleteLoreChapter
      tags: [lore-chapters]
      summary: Delete a chapter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Category:
      type: object
      required: [id, name, slug, description]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        articleCount:
          type: integer

    ArticleSummary:
      type: object
      required: [id, title, slug, categoryId, categoryName, excerpt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        excerpt:
          type: string
        updatedAt:
          type: string

    Article:
      type: object
      required: [id, title, slug, categoryId, categoryName, content, excerpt, updatedAt, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        parentId:
          type: integer
          nullable: true
        content:
          type: string
        excerpt:
          type: string
        updatedAt:
          type: string
        createdAt:
          type: string

    ArticleInput:
      type: object
      required: [title, slug, categoryId, content, excerpt]
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        categoryId:
          type: integer
        parentId:
          type: integer
          nullable: true
        content:
          type: string
        excerpt:
          type: string

    ArticleUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
        categoryId:
          type: integer
        parentId:
          type: integer
          nullable: true
        content:
          type: string
        excerpt:
          type: string

    Section:
      type: object
      required: [id, title, slug, description, link, icon, visible, sortOrder, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        link:
          type: string
        icon:
          type: string
        visible:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string

    SectionInput:
      type: object
      required: [title, slug]
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        description:
          type: string
        link:
          type: string
        icon:
          type: string
        visible:
          type: boolean
        sortOrder:
          type: integer

    SectionUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
        description:
          type: string
        link:
          type: string
        icon:
          type: string
        visible:
          type: boolean
        sortOrder:
          type: integer

    SectionOrder:
      type: object
      required: [order]
      properties:
        order:
          type: array
          items:
            type: integer
          description: Array of section IDs in desired order

    LoreChapter:
      type: object
      required: [id, title, content, sortOrder, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string

    LoreChapterInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        sortOrder:
          type: integer

    LoreChapterUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        sortOrder:
          type: integer

    ChapterOrder:
      type: object
      required: [order]
      properties:
        order:
          type: array
          items:
            type: integer
          description: Array of chapter IDs in desired order

    WikiStats:
      type: object
      required: [totalArticles, totalCategories, categoryBreakdown]
      properties:
        totalArticles:
          type: integer
        totalCategories:
          type: integer
        categoryBreakdown:
          type: array
          items:
            type: object
            required: [categoryName, count]
            properties:
              categoryName:
                type: string
              count:
                type: integer