# lib/api-zod/src/generated/types/articleInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Naruto RPG Community API
 * OpenAPI spec version: 0.1.0
 */

export interface ArticleInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  slug: string;
  categoryId: number;
  parentId?: number | null;
  content: string;
  excerpt: string;
}