# lib/api-zod/src/generated/types/articleSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Naruto RPG Community API
 * OpenAPI spec version: 0.1.0
 */

export interface ArticleSummary {
  id: number;
  title: string;
  slug: string;
  categoryId: number;
  categoryName: string;
  excerpt: string;
  updatedAt: string;
}