# lib/api-zod/src/generated/types/articleUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Naruto RPG Community API
 * OpenAPI spec version: 0.1.0
 */

export interface ArticleUpdate {
  /** @minLength 1 */
  title?: string;
  slug?: string;
  categoryId?: number;
  parentId?: number | null;
  content?: string;
  excerpt?: string;
}