# lib/api-zod/src/generated/types/loreChapterInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Naruto RPG Community API
 * OpenAPI spec version: 0.1.0
 */

export interface LoreChapterInput {
  /** @minLength 1 */
  title: string;
  content?: string;
  sortOrder?: number;
}