# lib/api-zod/src/generated/types/loreChapterUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Naruto RPG Community API
 * OpenAPI spec version: 0.1.0
 */

export interface LoreChapterUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  sortOrder?: number;
}