# lib/api-zod/src/generated/types/section.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Naruto RPG Community API
 * OpenAPI spec version: 0.1.0
 */

export interface Section {
  id: number;
  title: string;
  slug: string;
  description: string;
  link: string;
  icon: string;
  visible: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}