# lib/api-zod/src/generated/types/sectionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Naruto RPG Community API
 * OpenAPI spec version: 0.1.0
 */

export interface SectionInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  slug: string;
  description?: string;
  link?: string;
  icon?: string;
  visible?: boolean;
  sortOrder?: number;
}