# lib/api-zod/src/generated/types/sectionOrder.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Naruto RPG Community API
 * OpenAPI spec version: 0.1.0
 */

export interface SectionOrder {
  /** Array of section IDs in desired order */
  order: number[];
}