# lib/api-zod/src/generated/types/wikiStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Naruto RPG Community API
 * OpenAPI spec version: 0.1.0
 */
import type { WikiStatsCategoryBreakdownItem } from './wikiStatsCategoryBreakdownItem';

export interface WikiStats {
  totalArticles: number;
  totalCategories: number;
  categoryBreakdown: WikiStatsCategoryBreakdownItem[];
}